Deprecates the Reference Node
Adds @Depracated on GraphDatabaseService#getReferenceNode() and corresponding method in BatchInserter
Removes usage of the reference node in test cases that end up in the manual
Notes usage of the reference node in the ReST API for later removal
Changelog entry

Leaves usage of reference node in the shell. It is a larger change that needs some thought.