Fix for problems when using patterns with multiple unnamed nodes
To not have to use the same unnamed-numbering between parser versions, Pattern classes had a custom equals() method that ignored unnamed numbering.
This lead to a bug where patterns with multiple unnamed pattern relationships where dropped, and the user got a un-connected pattern exception.

The fix is to not ignore unnamed-numbers when doing equality checks. Unfortunately, this leads to having to have tests for the specific parsers, which is why
so many new tests where added to CypherParserTest