Removes object cache of nodes/relationships and their properties
Caching the entities read from (and written to) the storage layer has
served the purpose of speeding up a, comparatively slower store. Much have
changed over time in areas of hardware, knowledge of better access patterns
and a faster and more scalable storage layer have made the role of this
cache less and less important. On top of that caching is very complicated
and error prone and requires lots of memory to be effective, where garbage
collection of course gets more strained.

This commit removes everything that has to do with caching of nodes,
relationships and properties so that any reading of such will go directly
to the store (NeoStore, NodeStore, RelationshipStore a.s.o) without
anything in between. The Java API in org.neo4j.graphdb will not be
affected by this functionality-wise.

Performance for certain scenarios will be worse than with the cache.
This is expected and deemed OK for the time being. Removing this cache is
mainly an enabler for many future changes and optimizations with the goal
of surpassing the performance of neo4j with the cache. Key future goals
include: faster traversal, lower memory requirement, less garbage
collection, and not to forget, a simpler and smaller code base.

A memorial of classes deleted by this commit:
 - NodeImpl/RelationshipImpl: once upon a time one of the core parts of
   neo4j where relationship and property loading, iteration and caching
   took place
 - Cache(soft,weak,strong,none,hpc): took care of keeping references to
   NodeImpl/RelationshipImpl objects and adhering to memory limitations by
   evicting according to their own difference strategies
 - RelationshipLoader: loaded batches of relationships from store, caching
   into the requesting NodeImpl, in a RelIdArray
 - RelIdArray: responsible for caching of relationship lists per
   node/type/direction. Also provided means of iterating over those
   relationships
 - RelationshipLoadingPosition: keeping track of how much each cached
   NodeImpl had loaded in terms of relationships and where to start from
   when loading more