Solves problem with COST planner and ()-[*0..0]->() patterns
This is here to handle the *0..0 case. It means that both end nodes of the
relationship must be the same. Our solution to the problem is to keep count of
how many of these we see, and decrease the number of pattern nodes accordingly.
This workaround should work, but might not give the best numbers.

The nice solution would have been to rewrite these relationships away at an
earlier phase, but that is a much larger rewrite that we don't want to do in a
patch release.