Use relative paths for test data
For finding the initial test data files (import.cyp and params.json) getResource
is used in order to find the path relative to the classpath, rather than relative
to the current working directory of the JVM, since that will change depending on
how the test is executed. The CWD is different when running through maven than
when running through an IDE, whereas the classpath would contain the resource in
the same way in both cases.

For finding files referenced from the test data, the paths are instead relative
to the test data directory, i.e. the same directory that contains the import.cyp
and params.json files.