Better error handling when failing to satisfy all hints.
When the `greedy` planner end up in a state where it fails to satisfy all hints, it
should provide a better error message instead of just saying
`Failed to create a plan for the given QueryGraph`.

Note that when merging this into 2.3 we must change the test to explicitly
use `greedy` (since it is no longer the default).