Introduce Cypher SPI for code generation
With this SPI the Cypher compiler will know nothing about the underlying
methods it will be calling, only the operations it will want the generated
code to perform. The implementation of the SPI then knows about the specific
calls to make into the kernel API, generating code that makes those calls
directly.

The SPI is implemented through the use of the newly introduced code generation
library.