Fix unnecessary predicate exceptions
When we have multiple predicates AND'd together, if one evaluates to false
and another throws an exception, the exception is of no use, since the
entire row should not be returned. We evaluate these cases to false,
effectively ignoring the exception. If the other predicate was true, the
exception would of course be thrown as normal.