Add `dbms.security.load_csv_file_url_root` setting
Which restricts file URLs used with `LOAD CSV` to be within the
specified directory, such that the URL is resolved relative
to that root directory (similarly to a chroot).

There is currently no default for the setting, which allows file URLs to
access any local file. For neo4j-server, the default configuration files
explicitly set this to `data/import`, so new deployments will be
restricted by default. The `data/import` directory has also been added
to the server packaging.

For Neo4j 3.0, a default should be added to this property, so all
deployments are made safe by default.