Fix indeterministic test case in QueryInvalidationIT
The test for query replanning after adding data to populated database depended
on the randomization to succeed.
The new number of nodes created was not sufficient to exceed the default threshold
(since it is a strict > comparison), and the new random connections added could sometimes
fail to exceed the threshold if there were too many duplicates.

Now we calculate the number of nodes that need to be created based on the configuration.