Fix count query failures on empty database
Queries with simple count() expressions that can use the optimization of looking
up the count from the count store, and that counts nodes/relationships with
labels/relationship types, would fail if the label/relationship type did not exist in
the database (in particular with an empty database).
Now those counts should return 0.