Support shortestPath fallback on node-node case
The shortestPath algorithm returns directly when the start and end node are the same. However, this is not checked for predicate compliance like all other return cases, and leads to unexpected results, where predicates are not considered in this case. This commit adds that check and verifies that with predicates, the cypher fallback will be enacted even for same start and end nodes.