Revert changes around Supplier in Cypher compiler code generator
The problem is that the cypher-compiler-2.3 depends on
org.neo4j.function.Supplier and since 3.0 has a binary dependency
against the cypher-compiler-2.3 we cannot remove
org.neo4j.function.Supplier wihtout removing that dependency in 2.3.

Unfortunately, for the time being, we need to keep
org.neo4j.function.Supplier around.