Moved Storage Engine API back into kernel
Since kernel API and storage engine API are too interrelated to be separated.
Either both move or none do, they can always be broken out later. Having it
separately wasn't in vain though since it got the API cleaner and less leaky.

This also collapses the super tiny "neo4j-resource" component, which was really
a part of the graphdb-api back into kernel again. It was weird as a component,
but was temporarily necessary for having the storage engine API living outside kernel.

Even neo4j-proc (a tiny 4 class componentl) was moved into the kernel API
because it's a part of that API to begin with and could only really live on its own
due to the storage engine API was outside kernel.

This may seem like a step backwards, but things get a bit more coherent
for the time being and, again, this can be broken out later.