Fix eagerness bugs and add more tests
Fix bug when merge creates properties.
When a node pattern is matched with a property predicate, a later `MERGE` with
the same predicate should cause an eager plan to be inserted.

Fix bug where the read pattern and the write pattern uses the same nodes, but the
direction of the relationship is reversed.

- Add more eagerness acceptance tests
- Add separate assert parameter for optimal number of eagerness