Add a configuration option to disable shortestpath fallback
Shortestpath fallback enumerates all the paths in order to find the
shortest one, indeed in some cases graph algorithms cannot be used
(e.g., when using existential predicates on the path).  Since this
enumeration can be very expensive the user can decide to disable
fallback and the query will fail at runtime with an exception if the
graph algorithm is unable to find any path.

The default value for this config is false, which means that fallback
is enabled.