More efficient reuse of statement instances
Each KernelStatement instance is coupled with a KTI instances
and follows same initialize() calls as it. Likewise each StoreStatement
instance is coupled with a KernelStatement instance.

Starting a new transaction will extremely rarely create a new KTI instance,
otherwise most times pick one from a ThreadLocal with all these instances
already created, just waiting for a call to initialize() to be ready for use.
Likewise starting a statement just calls KernelStatement#acquire(), which in
turn calls StorageStatement#acquire(). Statement state is still cleated in
KernelStatement/StorageStatement#close(), but same instances will be reused
for the new statement (this was already done for KernelStatement). This commit
removes external pooling of StorageStatement instances since it's no longer
needed, so it's a net simplification of the code.