Use Vector as underlying type in results
When converting lists and sequence types to java we should make sure the
underlying type is an `IndexedSeq` so we get good performance for random
access methods (and `size`).

This problem was discovered in the context of procedures, where if you declare
your procedure to use a `List` you will always get a linked list back which
hurts performance for many common tasks.