Fix IDP solver configs
Several IDP solver configs were using only `expand` for solving the
given query graph.  The problem with this approach is that after the
first IDP iteration when we compress the IDP table we might end up
start expanding from another node which is not part of the pattern
solved so far.  In such a case not having joins will make IDP finding
no solutions.  In order to fix this problem we reenable `join`s after
the first compaction.