Assume single cardinality for empty/unknown statistics
When statistical data is unavailable or empty (such as in an empty
database), the cardinality becomes zero. This is a pitfall for cost
calculations, as the cardinality is a factor of the cost, and
multiplying by zero reduces the cost of any operation to zero. To still
be able to do reasonable cost comparisons, we assume the cardinality is
one for an empty/unknown label or pattern count.