Add per-procedure access mode capabilities
This commit introduces a new attribute to the `Procedure` annotation called `allowed`.
The value of that field is the name of a role in the security system, where any
user that is assigned to that role is given access to execute the procedure in
the procedure's specified access mode, regardless of any other access privileges
owned by any other roles that the user is assigned to.