Proper handle namespace resolution
It should be possible to use the names of the built-in functions as long
as you put the function in a separate namespace. The lookup was erroneously
using only the name of the function when doing lookup which lead to that
user-defined functions were shadowed by the built-in function with the same name.