Change unwind to use a WhileLoop instruction with iterators
Introduce two loop data generators that produces iterators for
collections of primitive vs. non-primitive types

This fixes an acceptance tests that failed because we were relying on
Iterable being the base type of the unwind collection expression
when type information was not known at compile time.

Move increment db hits from WhileLoop into the produceNext of the
loop data generators that are actually hitting the database.