Make Transformer generic in its context
This is a first step at making the compiler pipeline classes generic for use
with different state and context types. This step enables separating between
the compilation phases that live with the frontend and those that live with the
planner.

Context is split up into two pieces, one which has only dependencies on the frontend.