Change implementation of `any graph` to be nonempty
The `Given any graph` precondition is supposed to work on
exactly that: any graph. It is often mistaken for `empty graph`, however.
This commit makes `any graph` be a graph of 10 unlabeled nodes,
which did find three scenarios with a broken precondition. These
are now blacklisted while awaiting a TCK update.

Also added back ResultWrapper around results, to provide a better
error message when results do not match expectations.