Walking skeleton of OperationsFacade.indexQuery
The indexQuery() method accept an array of IndexQuery predicates, one for each
indexed property in order.
This API is able to descripte all of the other indexing operations in
OperationsFacade, though only seek and scan are suppported at this point.
The long term idea is to replace the other methods with this API, but keep the
QueryContext the same to maintain compatibility with other Cypher planner
versions.