Full pruning var length expand
This is an extended version of the Pruning Var Expand algorithm, improving
performance on paths with a max length of 5 or more, on randomized graphs.

The Full Pruning Var Expand (FullPrune) manages to prune paths below the min-length
by storing emit-state for all touched nodes, apart from the full expand depths. This
also makes the algorithm fully distinct, so no additional distinct operator is needed afterwards.