Unified SSL support and secure cluster communications
The SSL support of the neo4j platform has been extracted into a new SSL
module which allows configuring and defining SSL policies under distinct
policy names. These policies can then be utilized and shared by
different internal components that require SSL.

There is a legacy policy which derives from the old configuration
utilized by bolt and web servers which will be exposed by the new
module under the policy name of legacy. This legacy policy is however
exposed as an SslPolicy in the same way as policies defined in the
new system. The intention is to eventually deprecate the legacy
settings and their derived policy.

The purpose of this change is to unify and expand the configuration
surface and implementation code and allow for easy consumption of
SSL policies by various components existing today and in the future.

The new SSL policies allow plenty of flexibility, for example in
the allowed versions of SSL/TLS and ciphers. It also supports
setting up mutual authentication based on a PKI scheme.

The secure cluster communication builds on the SSL policy system
and works to secure Raft and TX channels between cluster members.

Some parts of the discovery channel communication implementation
is left as future work, hence it does not currently work.