InternalExecutionResult.executionPlanString() to avoid NPE when lifting
Lifting of previous versions execution Plans is not fully implemented.
A lot of cases simply pass null to the new objects of the plan
description. This leads to NullPointerExceptions when invoking toString.
Since the String is not used to assert but simply to show a user the
difference, we can avoid this by invoking toString on the non-lifted
plan description.

This cmooit also ports a couple of tests to the CypherComparisonSupport
and changes some of its internals slightly.

LdbcQueriesAcceptanceTest