Compute plan descriptions lazily
Turns out, computing the plan description is not free. It is
actually quite expensive. For simple queries like `MATCH (n) RETURN n.prop`,
we spend 10% of the time creating the plan description even though we rarely
need it. Changes so that plan descriptions are computed lazily when asked for
instead of on every executed query.