Make Property an extendable class
For the SlottedRewriter to be able to rewrite property expressions
where they are declared as Property, the corresponding runtime
expressions (e.g. NodeProperty) needs to extend Property.
To do this we change Property from a case class to a class, and
implement the methods that case class gave us to allow us to still
treat it as a case class (when matching, rewriting etc.).

Move calculation of needsExclusiveLock from the constructor of
the runtime SetOperation to the pipe builder, and compute it on the
AST expressions instead of the command expressions.