Unify pruning-var-expands
The FullPruningVarLength implementation is now optimized so it has similar
performance to the PruningVarLength implementation for short paths. It still
retains algorithmic performance benefits for long paths (about >= 4). Because
of this the old PruningVarLength has been decommissioned to test support (now
called LegacyPruningVarLengthExpandPipe), and the FullPruningLength is the
only production implementation (now just PruningVarLengthExpandPipe).