Non-recursive equals in LogicalPlan
LogicalPlan rewriting uses `fixedPoint` a lot which uses equality
to figure out if we reached a fix point. Logical plans are case classes
and use the default equal implementation meaning that we will recursively
compare objects over the logical plan tree. For large logical plan we
run out of call stack resulting in a stack overflow.