ActiveRead operator
This logical plan operator ensures that all source reads are performed on
the active tx layer. This means that any writes are immediately visible
without the need for a Transaction#markAsStable call, and is used in
the read part of merge plans to ensure that different rows in the same
merge clause see previous writes.