Implement ActiveReadPipe
ActiveReadPipe now modifies it's QueryState so that its query context reads
from the active tx-state layer. The modified QueryState is passed to the
source pipe#createResults call.

Note that we rely here on the kernel returning the same state Read both
on dataRead() and stableDataRead() if it doesn't support two-layer
tx-state. This is slightly suspicious, but was done because it seemed
painful to expose the kernel Modes all the way into Cypher.