Change TimeValue.writeTo and time() to use nanosPerDayLocal
We previously used nanosPerDayUTC, but this seemed less useful
as the value typically is specified in local time. To arrive
at a valid nanosPerDayUTC we might therefore overflow the valid
range as we add the offset to the nanosPerDayLocal, which would
a) require expensive modulus computations and b) mean that the
nanosPerDayUTC would not always compare as needed for TimeValue
comparison semantics. From a valid nanosPerDayLocal we can easily
compute an potentially invalid nanosPerDayUTC using the offset,
that can be used for semantically correct comparison.