Fix issue with cost planning index seeks
Make sure we get a minimum cardinality estimate of 1 when planning
leaf operators or filters, since it makes no sense to plan based on
estimations of 0.

A cardinality estimation of 0 on e.g. an index seek or scan is likely
to be caused by an empty label or incomplete or outdated statistics.
This can lead to a bad plan that does not scale.
And if the real cardinality turns out to indeed be 0,
it doesn't matter what plan we chose, so we might as well plan with
a safer estimate.