Add USING INDEX SEEK hint to Neo4j Cypher
The pre-existing hint USING INDEX can be fulfilled by both
index seek and index scan. USING INDEX SEEK follows the same syntax,
but is more specific and is only fulfilled by index seek.

This can be used in cases where the planner picks a plan with an
index scan based on lower cost estimation, that turns out to
perform badly in practice.